% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_online.R
\name{lda_svi}
\alias{lda_svi}
\title{Fit a Latent Dirichlet Allocation model to a text corpus}
\usage{
lda_svi(
  dtm,
  passes = 1,
  batchsize = 256,
  maxiter = 100,
  K,
  K0 = 0,
  eta = 1/K,
  alpha = 1/K,
  kappa = 0.7,
  tau_0 = 1024,
  refLambda = NULL
)
}
\arguments{
\item{dtm}{This may be a DocumentTermMatrix from the tm package, a simple_triplet_matrix from the slam package, or anything coercible to a simple_triplet_matrix. This includes a data frame with the document names in the first column, the terms in the second column, and the number or appearances in the third column.}

\item{passes}{How many times we look at each document.}

\item{batchsize}{The size of the minibatches.}

\item{maxiter}{The maximum iterations for the "E step" for each document (the updating of the per-document parameters within each minibatch). The default of 100 follows the reference implementation in python by the authors.}

\item{K}{The number of topics}

\item{eta}{hyperparameter for the term}

\item{alpha}{hyperparameter.}

\item{kappa}{learning rate parameter. Lower values give greater weight to later iterations. For guaranteed convergence to a local optimum, kappa must lie in the interval (0.5,1].}

\item{tau_0}{learning rate parameter. Higher values reduce the influence of early iterations.}
}
\description{
Fit a Latent Dirichlet Allocation model to a text corpus
}
