% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.svi-package.R
\docType{package}
\name{RefTM-package}
\alias{RefTM}
\alias{RefTM-package}
\title{RefTM: Reference-guided Topic Modeling of scCAS data}
\description{
RefTM is a R package for the integrative analysis of scCAS data.
}
\keyword{internal}
